uses PROGRAM mode, fixed function defs and use

+++ QB/mark.h
@@ -2,6 +2,6 @@
 #include "mode.h"
 #include "pquestions.h"
 
-extern int question_correct(char*,uint32_t,int,int,char*);
+extern int question_correct(char*,int64_t,int,int,char*);
 extern char *PY_A; //ans for python
 extern char *C_A; //ans for c
